Add glslify
Turns out that the "asset/source" bit just above the glslify
loader was already doing the work raw-loader would be doing,
and so all we had to do was not include raw-loader :)

One side effect of adding this is that the final output from
glslify is shown on the page, rather than the original source.
Sometimes this can be a bit messy, depending on the modules
you're using and how they're being used.